requirements/mypy.txt
mypy==1.8.0
mypy-zope==1.0.3
types-click==7.1.8
